worker: extract worker-side CoS helpers into worker/cos.rs (#957 P1)
Move the worker-side CoS runtime helpers + their tests out of
worker.rs (now worker/mod.rs) into a sibling worker/cos.rs.
14 production fns + 21 tests + 3 test helpers move:

Production:
- build_worker_cos_owner_live_by_tx_ifindex
- queue_uses_shared_exact_service
- build_worker_cos_fast_interfaces / build_worker_cos_statuses /
  build_worker_cos_statuses_from_maps
- unique_owner_profile_row, cos_runtime_config_changed
- vacate_all_shared_exact_slots_for_binding
- reset_binding_cos_runtime / reset_worker_cos_runtimes
- owner_profile_snapshot
- merge_cos_queue_owner_profile_sum / merge_owner_profile_sum /
  merge_binding_scoped_owner_profile

Tests: the entire `#[cfg(test)] mod tests` block moves with the
production code (per modularity-discipline test-colocation rule).
All 21 tests in the block were CoS-specific.

Mechanical:
- worker.rs → worker/mod.rs (directory module)
- afxdp.rs `#[path]` updated to point at the new mod.rs
- worker/mod.rs declares `mod cos;` and imports the helpers via
  explicit `use cos::{...}` lists (matches checksum.rs / inspect.rs
  precedent for re-exports)
- The pre-existing `use super::cos::{...}` import (from afxdp::cos
  directory module) is unchanged — `super` from worker/mod.rs is
  still afxdp; the new `worker::cos` is at a different path

worker/mod.rs production LOC (excluding mod tests):
  2,608 → 1,938 — UNDER the 2,000 modularity-discipline threshold
worker/cos.rs: 704 prod LOC (well under threshold)

Visibility: fns kept at lowest-needed level
- pub(super) for items used only inside worker/
- pub(in crate::afxdp) for items already at that visibility before
  the move (owner_profile_snapshot, merge_owner_profile_sum,
  merge_binding_scoped_owner_profile)
- pub(crate) for merge_cos_queue_owner_profile_sum (was already
  pub(crate))

Per docs/refactor/957 (modularize afxdp.rs and worker.rs core loops).
afxdp.rs is already at 1,051 prod LOC, so this PR addresses the
worker.rs half.

build clean (cargo build --release), 865 Rust tests pass.
