+++ userspace-dp/src/afxdp/worker/mod.rs
@@ -1,0 +1,1938 @@
+use super::*;
+
+// #957 P1: worker-side CoS runtime helpers split out into a sibling
+// submodule. Note this module is `worker::cos`, separate from the
+// `afxdp::cos` directory module imported below as `super::cos`.
+mod cos;
+use cos::{
+    build_worker_cos_fast_interfaces, build_worker_cos_owner_live_by_tx_ifindex,
+    build_worker_cos_statuses, cos_runtime_config_changed, reset_binding_cos_runtime,
+    reset_worker_cos_runtimes, vacate_all_shared_exact_slots_for_binding,
+};
+pub(crate) use cos::merge_cos_queue_owner_profile_sum;
+pub(super) use cos::{
+    merge_binding_scoped_owner_profile, merge_owner_profile_sum, owner_profile_snapshot,
+};
+
+// #956 Phase 4-5: explicit imports for items that moved out of tx.rs into
+// cos/token_bucket.rs (Phase 4) and cos/queue_ops.rs (Phase 5). Without
+// this, neither the local `use super::*;` glob nor afxdp.rs's
+// `use self::tx::*;` parent-module glob still reaches them — the items
+// no longer originate from tx.rs after the moves.
+use super::cos::{
+    cos_queue_len, cos_queue_pop_front_no_snapshot, release_all_cos_queue_leases,
+    release_all_cos_root_leases,
+};
+
+pub(crate) struct BindingWorker {
+    pub(crate) slot: u32,
+    pub(crate) queue_id: u32,
+    pub(crate) worker_id: u32,
+    pub(crate) interface: Arc<str>,
+    pub(crate) ifindex: i32,
+    pub(crate) umem: WorkerUmem,
+    pub(crate) live: Arc<BindingLiveState>,
+    #[allow(dead_code)]
+    pub(crate) user: User,
+    pub(crate) device: crate::xsk_ffi::DeviceQueue,
+    pub(crate) rx: crate::xsk_ffi::RingRx,
+    pub(crate) tx: crate::xsk_ffi::RingTx,
+    pub(crate) free_tx_frames: VecDeque<u64>,
+    pub(crate) pending_tx_prepared: VecDeque<PreparedTxRequest>,
+    pub(crate) pending_tx_local: VecDeque<TxRequest>,
+    pub(crate) max_pending_tx: usize,
+    pub(crate) cos_fast_interfaces: FastMap<i32, WorkerCoSInterfaceFastPath>,
+    pub(crate) cos_interfaces: FastMap<i32, CoSInterfaceRuntime>,
+    pub(crate) cos_interface_order: Vec<i32>,
+    pub(crate) cos_interface_rr: usize,
+    pub(crate) cos_nonempty_interfaces: usize,
+    pub(crate) pending_fill_frames: VecDeque<u64>,
+    pub(crate) scratch_recycle: Vec<u64>,
+    pub(crate) scratch_forwards: Vec<PendingForwardRequest>,
+    pub(crate) scratch_fill: Vec<u64>,
+    pub(crate) scratch_prepared_tx: Vec<PreparedTxRequest>,
+    pub(crate) scratch_local_tx: Vec<(u64, TxRequest)>,
+    pub(crate) scratch_exact_prepared_tx: Vec<ExactPreparedScratchTxRequest>,
+    pub(crate) scratch_exact_local_tx: Vec<ExactLocalScratchTxRequest>,
+    pub(crate) scratch_completed_offsets: Vec<u64>,
+    pub(crate) scratch_post_recycles: Vec<(u32, u64)>,
+    /// #812: per-UMEM-frame submit timestamp sidecar. Indexed by
+    /// `offset >> UMEM_FRAME_SHIFT`. Pre-allocated once at binding
+    /// construction (length = total UMEM frames) so the hot-path
+    /// stamp write is a single store — NO allocation, NO grow.
+    ///
+    /// Single-writer invariant (plan §3.3): submit and completion for
+    /// one frame offset are the same thread (the owner worker that
+    /// owns this binding's `free_tx_frames` via
+    /// `pop_front`/`push_front`), so plain `Vec<u64>` is correct —
+    /// no atomic needed. `WorkerUmem` is `Rc` (not `Arc`) at
+    /// `umem.rs:16-18`, enforcing single-owner semantics even when
+    /// multiple bindings share a UMEM under the mlx5 special case.
+    ///
+    /// Unstamped slots hold `TX_SIDECAR_UNSTAMPED` (`u64::MAX`) — the
+    /// reap path (`reap_tx_completions`) skips the histogram
+    /// increment for these to avoid biasing the tail toward bucket 0
+    /// (plan §5.4). A `monotonic_nanos() == 0` return (VDSO failure,
+    /// plan §3.4a / §6.1 test #5) causes `stamp_submits` to early-
+    /// return without writing the slot (Codex round-1 MED + Rust
+    /// round-1 MED-2) — the slot's pre-existing UNSTAMPED state
+    /// (from the previous reap or from worker construction) is what
+    /// the reap checks.
+    /// Rust round-1 MED-1: `Box<[u64]>` rather than `Vec<u64>` to
+    /// convey single-size intent at the type level. The sidecar is
+    /// pre-allocated to `total_frames` at binding construction and
+    /// NEVER grown. A future refactor that naively added `push` to
+    /// `Vec<u64>` would silently allocate on the hot path; `Box<[u64]>`
+    /// has no `push` method, so the mistake fails to compile.
+    pub(crate) tx_submit_ns: Box<[u64]>,
+    /// Packets waiting for neighbor resolution. The UMEM frame is held
+    /// (not recycled) until the neighbor resolves or the entry times out.
+    pub(crate) pending_neigh: VecDeque<PendingNeighPacket>,
+    /// Flow cache fast-path: cross-binding in-place rewrites deferred
+    /// until after the RX batch (borrow checker prevents mutable access
+    /// to two bindings simultaneously inside the RX loop).
+    #[allow(dead_code)] // reserved for cross-binding fast-path
+    pub(crate) scratch_cross_binding_tx: Vec<(usize, PreparedTxRequest)>,
+    pub(crate) scratch_rst_teardowns: Vec<(SessionKey, NatDecision)>,
+    pub(crate) in_flight_prepared_recycles: FastMap<u64, PreparedTxRecycle>,
+    pub(crate) heartbeat_map_fd: c_int,
+    pub(crate) session_map_fd: c_int,
+    pub(crate) conntrack_v4_fd: c_int,
+    pub(crate) conntrack_v6_fd: c_int,
+    pub(crate) last_heartbeat_update_ns: u64,
+    pub(crate) debug_state_counter: u32,
+    pub(crate) last_rx_wake_ns: u64,
+    pub(crate) last_tx_wake_ns: u64,
+    pub(crate) outstanding_tx: u32,
+    pub(crate) empty_rx_polls: u32,
+    pub(crate) last_learned_neighbor: Option<LearnedNeighborKey>,
+    pub(crate) dbg_fill_submitted: u64,
+    pub(crate) dbg_fill_failed: u64,
+    pub(crate) dbg_poll_cycles: u64,
+    pub(crate) dbg_backpressure: u64,
+    pub(crate) dbg_rx_empty: u64,
+    pub(crate) dbg_rx_wakeups: u64,
+    // TX pipeline debug counters
+    pub(crate) dbg_tx_ring_submitted: u64, // descriptors inserted into TX ring
+    pub(crate) dbg_tx_ring_full: u64,      // times TX ring insert returned 0
+    pub(crate) dbg_completions_reaped: u64, // completion descriptors read
+    pub(crate) dbg_sendto_calls: u64,      // number of sendto/wake calls
+    pub(crate) dbg_sendto_err: u64,        // sendto returned error (non-EAGAIN/ENOBUFS)
+    pub(crate) dbg_sendto_eagain: u64,     // sendto returned EAGAIN/EWOULDBLOCK
+    pub(crate) dbg_sendto_enobufs: u64,    // sendto returned ENOBUFS (kernel TX drop)
+    // #802/#804: per-binding bound-pending FIFO overflow — incremented at
+    // the `bound_pending_tx_local`/`bound_pending_tx_prepared` evict
+    // sites only. Pre-#804 this counter was named `dbg_pending_overflow`
+    // and also doubled as the CoS admission overflow accumulator; the
+    // two semantics are now tracked on separate fields so operators can
+    // tell which path is dropping.
+    pub(crate) dbg_bound_pending_overflow: u64,
+    // #804: per-binding class-of-service queue admission overflow —
+    // incremented in `enqueue_cos_item()` when the CoS admission gate
+    // rejects the item.
+    pub(crate) dbg_cos_queue_overflow: u64,
+    pub(crate) dbg_tx_tcp_rst: u64,        // TCP RST packets transmitted
+    // Ring diagnostics — raw values from xsk_ffi API
+    pub(crate) dbg_rx_avail_nonzero: u64, // times rx.available() > 0
+    pub(crate) dbg_rx_avail_max: u32,     // max rx.available() seen this interval
+    pub(crate) dbg_fill_pending: u32,     // fill ring: userspace produced - kernel consumed
+    pub(crate) dbg_device_avail: u32,     // device queue available (completion ring pending)
+    pub(crate) dbg_rx_wake_sendto_ok: u64, // sendto() returned >= 0 in maybe_wake_rx
+    pub(crate) dbg_rx_wake_sendto_err: u64, // sendto() returned < 0 in maybe_wake_rx
+    pub(crate) dbg_rx_wake_sendto_errno: i32, // last errno from sendto in maybe_wake_rx
+    pub(crate) pending_direct_tx_packets: u64,
+    pub(crate) pending_copy_tx_packets: u64,
+    pub(crate) pending_in_place_tx_packets: u64,
+    pub(crate) pending_direct_tx_no_frame_fallback_packets: u64,
+    pub(crate) pending_direct_tx_build_fallback_packets: u64,
+    pub(crate) pending_direct_tx_disallowed_fallback_packets: u64,
+    pub(crate) flow_cache: FlowCache,
+    pub(crate) flow_cache_session_touch: u64,
+    /// Timestamp when this binding was created.
+    #[allow(dead_code)] // reserved for heartbeat gating logic
+    pub(crate) bind_time_ns: u64,
+    /// Zero-copy vs copy mode (affects heartbeat gating).
+    #[allow(dead_code)] // reserved for heartbeat gating logic
+    pub(crate) bind_mode: XskBindMode,
+    /// Set true once the XSK RX ring has delivered at least one packet,
+    /// proving the NIC's XSK receive queue is active for this binding.
+    pub(crate) xsk_rx_confirmed: bool,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub(crate) enum XskBindMode {
+    Unknown,
+    Copy,
+    ZeroCopy,
+}
+
+impl XskBindMode {
+    pub(crate) fn as_u8(self) -> u8 {
+        match self {
+            Self::Unknown => 0,
+            Self::Copy => 1,
+            Self::ZeroCopy => 2,
+        }
+    }
+
+    pub(crate) fn from_u8(value: u8) -> Self {
+        match value {
+            1 => Self::Copy,
+            2 => Self::ZeroCopy,
+            _ => Self::Unknown,
+        }
+    }
+
+    pub(crate) fn as_str(self) -> &'static str {
+        match self {
+            Self::Unknown => "",
+            Self::Copy => "copy",
+            Self::ZeroCopy => "zerocopy",
+        }
+    }
+
+    pub(crate) fn is_zerocopy(self) -> bool {
+        matches!(self, Self::ZeroCopy)
+    }
+}
+
+pub(crate) fn fabric_queue_hash(
+    flow: Option<&SessionFlow>,
+    expected_ports: Option<(u16, u16)>,
+    meta: UserspaceDpMeta,
+) -> u64 {
+    fn mix(seed: &mut u64, value: u64) {
+        *seed ^= value
+            .wrapping_add(0x9e3779b97f4a7c15)
+            .wrapping_add(*seed << 6)
+            .wrapping_add(*seed >> 2);
+    }
+
+    let mut seed = meta.protocol as u64;
+    if let Some(flow) = flow {
+        match flow.src_ip {
+            IpAddr::V4(ip) => mix(&mut seed, u32::from(ip) as u64),
+            IpAddr::V6(ip) => {
+                for chunk in ip.octets().chunks_exact(8) {
+                    mix(&mut seed, u64::from_be_bytes(chunk.try_into().unwrap()));
+                }
+            }
+        }
+        match flow.dst_ip {
+            IpAddr::V4(ip) => mix(&mut seed, u32::from(ip) as u64),
+            IpAddr::V6(ip) => {
+                for chunk in ip.octets().chunks_exact(8) {
+                    mix(&mut seed, u64::from_be_bytes(chunk.try_into().unwrap()));
+                }
+            }
+        }
+        mix(&mut seed, flow.forward_key.src_port as u64);
+        mix(&mut seed, flow.forward_key.dst_port as u64);
+        return seed;
+    }
+    let (src_port, dst_port) = expected_ports.unwrap_or((meta.flow_src_port, meta.flow_dst_port));
+    mix(&mut seed, src_port as u64);
+    mix(&mut seed, dst_port as u64);
+    seed
+}
+
+#[derive(Clone, Debug)]
+pub(crate) struct SyncedSessionEntry {
+    pub(crate) key: SessionKey,
+    pub(crate) decision: SessionDecision,
+    pub(crate) metadata: SessionMetadata,
+    pub(crate) origin: SessionOrigin,
+    pub(crate) protocol: u8,
+    pub(crate) tcp_flags: u8,
+}
+
+impl BindingWorker {
+    fn create(
+        binding: &BindingStatus,
+        ring_entries: u32,
+        xsk_map_fd: c_int,
+        heartbeat_map_fd: c_int,
+        session_map_fd: c_int,
+        conntrack_v4_fd: c_int,
+        conntrack_v6_fd: c_int,
+        live: Arc<BindingLiveState>,
+        bind_strategy: AfXdpBindStrategy,
+        poll_mode: crate::PollMode,
+        mut worker_umem: WorkerUmem,
+        frame_pool: &mut VecDeque<u64>,
+        shared_umem: bool,
+    ) -> Result<Self, Box<dyn std::error::Error + Send + Sync>> {
+        let driver_name = interface_driver_name(&binding.interface);
+        let total_frames =
+            binding_frame_count_for_driver(driver_name.as_deref(), ring_entries).max(1);
+        let reserved_tx =
+            reserved_tx_frames_for_driver(driver_name.as_deref(), ring_entries).min(total_frames);
+        let mut reserved_tx_frames = VecDeque::with_capacity(reserved_tx as usize);
+        for _ in 0..reserved_tx {
+            let Some(offset) = frame_pool.pop_front() else {
+                return Err(format!(
+                    "insufficient shared UMEM frames for reserved TX on {} if{}q{}",
+                    binding.interface, binding.ifindex, binding.queue_id
+                )
+                .into());
+            };
+            reserved_tx_frames.push_back(offset);
+        }
+        // Pre-populate fill ring with ALL remaining frames — no spare held back.
+        // This maximizes the kernel's ability to place received packets and
+        // prevents fill ring starvation under burst conditions (copy-mode fix).
+        let mut initial_fill_frames = Vec::with_capacity((total_frames - reserved_tx) as usize);
+        for _ in reserved_tx..total_frames {
+            let Some(offset) = frame_pool.pop_front() else {
+                return Err(format!(
+                    "insufficient shared UMEM frames for fill ring on {} if{}q{}",
+                    binding.interface, binding.ifindex, binding.queue_id
+                )
+                .into());
+            };
+            initial_fill_frames.push(offset);
+        }
+        let info = ifinfo_from_binding(binding)?;
+        let (user, rx, tx, bind_mode, actual_bind_strategy, device) = open_binding_worker_rings(
+            &mut worker_umem,
+            &info,
+            ring_entries,
+            bind_strategy,
+            driver_name.as_deref(),
+            poll_mode,
+            Some(&initial_fill_frames),
+        )
+        .map_err(|err| format!("configure AF_XDP rings: {err}"))?;
+
+        let user_fd = user.as_raw_fd();
+        live.set_bound(user_fd);
+        live.set_bind_mode(bind_mode);
+        // getsockname() returns ENOTSUP on AF_XDP sockets (kernel doesn't
+        // implement it for this family).  Use the binding plan's expected
+        // ifindex/queue_id directly — umem.bind() already validated these.
+        live.set_socket_binding(binding.ifindex, binding.queue_id, 0);
+        // #878: publish per-binding capacities so the snapshot path can
+        // expose them via the wire BindingStatus. These are write-once
+        // (set here at worker construction) and read-many.
+        live.umem_total_frames
+            .store(total_frames, std::sync::atomic::Ordering::Relaxed);
+        live.tx_ring_capacity
+            .store(ring_entries, std::sync::atomic::Ordering::Relaxed);
+        eprintln!(
+            "xpf-userspace-dp: binding slot={} fd={} strategy={} bound if{}q{} mode={:?} shared_umem={}",
+            binding.slot,
+            user_fd,
+            actual_bind_strategy.describe(),
+            binding.ifindex,
+            binding.queue_id,
+            bind_mode,
+            shared_umem,
+        );
+        if let Err(err) = register_xsk_slot(xsk_map_fd, binding.slot, user_fd) {
+            eprintln!(
+                "xpf-userspace-dp: ERROR register_xsk_slot slot={} fd={}: {}",
+                binding.slot, user_fd, err,
+            );
+            live.set_error(format!("register XSK slot: {err}"));
+        } else {
+            eprintln!(
+                "xpf-userspace-dp: registered slot={} fd={} in XSKMAP",
+                binding.slot, user_fd,
+            );
+            live.set_xsk_registered(true);
+            live.clear_error();
+        }
+        let init_now = monotonic_nanos();
+        let max_pending_tx = pending_tx_capacity(ring_entries);
+        if let Err(err) = touch_heartbeat(heartbeat_map_fd, binding.slot, &live, init_now) {
+            live.set_error(format!("update heartbeat slot: {err}"));
+        }
+        live.set_max_pending_tx(max_pending_tx);
+        let mut binding = Self {
+            slot: binding.slot,
+            queue_id: binding.queue_id,
+            worker_id: binding.worker_id,
+            interface: Arc::<str>::from(binding.interface.as_str()),
+            ifindex: binding.ifindex,
+            umem: worker_umem,
+            live,
+            user,
+            device,
+            rx,
+            tx,
+            free_tx_frames: reserved_tx_frames,
+            pending_tx_prepared: VecDeque::new(),
+            pending_tx_local: VecDeque::new(),
+            max_pending_tx,
+            cos_fast_interfaces: FastMap::default(),
+            cos_interfaces: FastMap::default(),
+            cos_interface_order: Vec::new(),
+            cos_interface_rr: 0,
+            cos_nonempty_interfaces: 0,
+            pending_fill_frames: VecDeque::new(),
+            scratch_recycle: Vec::with_capacity(RX_BATCH_SIZE as usize),
+            scratch_forwards: Vec::with_capacity(RX_BATCH_SIZE as usize),
+            scratch_fill: Vec::with_capacity(FILL_BATCH_SIZE),
+            scratch_prepared_tx: Vec::with_capacity(TX_BATCH_SIZE),
+            scratch_local_tx: Vec::with_capacity(TX_BATCH_SIZE),
+            scratch_exact_prepared_tx: Vec::with_capacity(TX_BATCH_SIZE),
+            scratch_exact_local_tx: Vec::with_capacity(TX_BATCH_SIZE),
+            scratch_completed_offsets: Vec::with_capacity(ring_entries as usize),
+            scratch_post_recycles: Vec::with_capacity(RX_BATCH_SIZE as usize),
+            // #812: pre-allocate the submit-timestamp sidecar once,
+            // sized to the binding's total UMEM frame count so every
+            // legal `offset >> UMEM_FRAME_SHIFT` index lands inside
+            // the vec. Initial contents are the unstamped sentinel so
+            // any stray pre-existing offset in flight (cross-restart
+            // completion) is skipped by the reap path (plan §5.4).
+            // Allocation happens here — NEVER on the hot path.
+            // Rust round-1 MED-1: Box<[u64]> — allocate-once, never
+            // grow. `vec![...].into_boxed_slice()` produces an
+            // exactly-sized heap allocation with no spare capacity.
+            tx_submit_ns: vec![TX_SIDECAR_UNSTAMPED; total_frames as usize].into_boxed_slice(),
+            pending_neigh: VecDeque::with_capacity(MAX_PENDING_NEIGH),
+            scratch_cross_binding_tx: Vec::with_capacity(RX_BATCH_SIZE as usize),
+            scratch_rst_teardowns: Vec::with_capacity(16),
+            in_flight_prepared_recycles: FastMap::default(),
+            heartbeat_map_fd,
+            session_map_fd,
+            conntrack_v4_fd,
+            conntrack_v6_fd,
+            last_heartbeat_update_ns: init_now,
+            debug_state_counter: 0,
+            last_rx_wake_ns: init_now,
+            last_tx_wake_ns: init_now,
+            outstanding_tx: 0,
+            empty_rx_polls: 0,
+            last_learned_neighbor: None,
+            dbg_fill_submitted: 0,
+            dbg_fill_failed: 0,
+            dbg_poll_cycles: 0,
+            dbg_backpressure: 0,
+            dbg_rx_empty: 0,
+            dbg_rx_wakeups: 0,
+            dbg_tx_ring_submitted: 0,
+            dbg_tx_ring_full: 0,
+            dbg_completions_reaped: 0,
+            dbg_sendto_calls: 0,
+            dbg_sendto_err: 0,
+            dbg_sendto_eagain: 0,
+            dbg_sendto_enobufs: 0,
+            dbg_bound_pending_overflow: 0,
+            dbg_cos_queue_overflow: 0,
+            dbg_tx_tcp_rst: 0,
+            dbg_rx_avail_nonzero: 0,
+            dbg_rx_avail_max: 0,
+            dbg_fill_pending: 0,
+            dbg_device_avail: 0,
+            dbg_rx_wake_sendto_ok: 0,
+            dbg_rx_wake_sendto_err: 0,
+            dbg_rx_wake_sendto_errno: 0,
+            pending_direct_tx_packets: 0,
+            pending_copy_tx_packets: 0,
+            pending_in_place_tx_packets: 0,
+            pending_direct_tx_no_frame_fallback_packets: 0,
+            pending_direct_tx_build_fallback_packets: 0,
+            pending_direct_tx_disallowed_fallback_packets: 0,
+            flow_cache: FlowCache::new(),
+            flow_cache_session_touch: 0,
+            bind_time_ns: {
+                let mut ts = libc::timespec {
+                    tv_sec: 0,
+                    tv_nsec: 0,
+                };
+                unsafe { libc::clock_gettime(libc::CLOCK_MONOTONIC, &mut ts) };
+                ts.tv_sec as u64 * 1_000_000_000 + ts.tv_nsec as u64
+            },
+            xsk_rx_confirmed: false,
+            bind_mode,
+        };
+        update_binding_debug_state(&mut binding);
+        Ok(binding)
+    }
+
+    pub(crate) fn identity(&self) -> BindingIdentity {
+        BindingIdentity {
+            slot: self.slot,
+            queue_id: self.queue_id,
+            worker_id: self.worker_id,
+            interface: self.interface.clone(),
+            ifindex: self.ifindex,
+        }
+    }
+}
+pub(crate) fn worker_loop(
+    worker_id: u32,
+    binding_plans: Vec<BindingPlan>,
+    shared_validation: Arc<ArcSwap<ValidationState>>,
+    shared_forwarding: Arc<ArcSwap<ForwardingState>>,
+    ha_state: Arc<ArcSwap<BTreeMap<i32, HAGroupRuntime>>>,
+    dynamic_neighbors: Arc<ShardedNeighborMap>,
+    shared_sessions: Arc<Mutex<FastMap<SessionKey, SyncedSessionEntry>>>,
+    shared_nat_sessions: Arc<Mutex<FastMap<SessionKey, SyncedSessionEntry>>>,
+    shared_forward_wire_sessions: Arc<Mutex<FastMap<SessionKey, SyncedSessionEntry>>>,
+    shared_owner_rg_indexes: SharedSessionOwnerRgIndexes,
+    slow_path: Option<Arc<SlowPathReinjector>>,
+    local_tunnel_deliveries: Arc<ArcSwap<BTreeMap<i32, SyncSender<Vec<u8>>>>>,
+    recent_exceptions: Arc<Mutex<VecDeque<ExceptionStatus>>>,
+    recent_session_deltas: Arc<Mutex<VecDeque<SessionDeltaInfo>>>,
+    last_resolution: Arc<Mutex<Option<PacketResolution>>>,
+    commands: Arc<Mutex<VecDeque<WorkerCommand>>>,
+    peer_worker_commands: Vec<Arc<Mutex<VecDeque<WorkerCommand>>>>,
+    worker_commands_by_id: Arc<BTreeMap<u32, Arc<Mutex<VecDeque<WorkerCommand>>>>>,
+    stop: Arc<AtomicBool>,
+    heartbeat: Arc<AtomicU64>,
+    session_export_ack: Arc<AtomicU64>,
+    poll_mode: crate::PollMode,
+    dnat_fds: DnatTableFds,
+    shared_fabrics: Arc<ArcSwap<Vec<FabricLink>>>,
+    event_stream: Option<crate::event_stream::EventStreamWorkerHandle>,
+    rg_epochs: Arc<[AtomicU32; MAX_RG_EPOCHS]>,
+    shared_cos_owner_worker_by_queue: Arc<ArcSwap<BTreeMap<(i32, u8), u32>>>,
+    shared_cos_owner_live_by_queue: Arc<ArcSwap<BTreeMap<(i32, u8), Arc<BindingLiveState>>>>,
+    shared_cos_root_leases: Arc<ArcSwap<BTreeMap<i32, Arc<SharedCoSRootLease>>>>,
+    shared_cos_queue_leases: Arc<ArcSwap<BTreeMap<(i32, u8), Arc<SharedCoSQueueLease>>>>,
+    shared_cos_queue_vtime_floors: Arc<
+        ArcSwap<BTreeMap<(i32, u8), Arc<SharedCoSQueueVtimeFloor>>>,
+    >,
+    cos_status: Arc<ArcSwap<Vec<crate::protocol::CoSInterfaceStatus>>>,
+    // #869: worker-runtime telemetry publish slot.  Worker writes its
+    // local counters here on a ~1s cadence; coordinator reads for status.
+    runtime_atomics: Arc<super::worker_runtime::WorkerRuntimeAtomics>,
+) {
+    pin_current_thread(worker_id);
+    const COS_STATUS_INTERVAL_NS: u64 = 100_000_000;
+    let ha_startup_grace_until_secs =
+        (monotonic_nanos() / 1_000_000_000).saturating_add(TUNNEL_HA_STARTUP_GRACE_SECS);
+    let mut validation = **shared_validation.load();
+    let mut forwarding = shared_forwarding.load_full();
+    let mut cos_owner_worker_by_queue = shared_cos_owner_worker_by_queue.load_full();
+    let mut cos_owner_live_by_queue = shared_cos_owner_live_by_queue.load_full();
+    let mut cos_shared_root_leases = shared_cos_root_leases.load_full();
+    let mut cos_shared_queue_leases = shared_cos_queue_leases.load_full();
+    let mut cos_shared_queue_vtime_floors = shared_cos_queue_vtime_floors.load_full();
+    let mut sessions = SessionTable::new();
+    let mut screen_state = ScreenState::new();
+    screen_state.update_profiles(forwarding.screen_profiles.clone());
+    sessions.set_timeouts(forwarding.session_timeouts);
+    let mut bindings = Vec::with_capacity(binding_plans.len());
+    for plan in binding_plans {
+        let driver_name = interface_driver_name(&plan.status.interface);
+        let total_frames =
+            binding_frame_count_for_driver(driver_name.as_deref(), plan.ring_entries).max(1);
+        let binding = match WorkerUmemPool::new(total_frames)
+            .map_err(|err| format!("create binding umem: {err}"))
+        {
+            Ok(WorkerUmemPool {
+                umem,
+                mut free_frames,
+            }) => BindingWorker::create(
+                &plan.status,
+                plan.ring_entries,
+                plan.xsk_map_fd,
+                plan.heartbeat_map_fd,
+                plan.session_map_fd,
+                plan.conntrack_v4_fd,
+                plan.conntrack_v6_fd,
+                plan.live.clone(),
+                plan.bind_strategy,
+                plan.poll_mode,
+                umem,
+                &mut free_frames,
+                false,
+            ),
+            Err(err) => Err(err.to_string().into()),
+        };
+        match binding {
+            Ok(binding) => bindings.push(binding),
+            Err(err) => plan.live.set_error(err.to_string()),
+        }
+    }
+    let binding_lookup = WorkerBindingLookup::from_bindings(&bindings);
+    let cos_owner_live_by_tx_ifindex = build_worker_cos_owner_live_by_tx_ifindex(
+        bindings
+            .iter()
+            .map(|binding| (binding.ifindex, binding.live.clone())),
+    );
+    let cos_fast_interfaces = build_worker_cos_fast_interfaces(
+        forwarding.as_ref(),
+        worker_id,
+        &cos_owner_live_by_tx_ifindex,
+        cos_owner_worker_by_queue.as_ref(),
+        cos_owner_live_by_queue.as_ref(),
+        cos_shared_root_leases.as_ref(),
+        cos_shared_queue_leases.as_ref(),
+        cos_shared_queue_vtime_floors.as_ref(),
+    );
+    for binding in bindings.iter_mut() {
+        binding.cos_fast_interfaces = cos_fast_interfaces.clone();
+    }
+    let mut interrupt_poll_fds = if poll_mode == crate::PollMode::Interrupt {
+        bindings
+            .iter()
+            .map(|binding| libc::pollfd {
+                fd: binding.device.as_raw_fd(),
+                events: libc::POLLIN,
+                revents: 0,
+            })
+            .collect::<Vec<_>>()
+    } else {
+        Vec::new()
+    };
+    let mut idle_iters = 0u32;
+    let mut poll_start = 0usize;
+    let mut shared_recycles = Vec::with_capacity((RX_BATCH_SIZE as usize).saturating_mul(2));
+    // Debug: periodic summary counters
+    let mut dbg_last_report_ns = monotonic_nanos();
+    let mut dbg_rx_total = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_tx_total = 0u64;
+    let mut dbg_forward_total = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_local_total = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_session_hit = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_session_miss = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_session_create = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_no_route = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_missing_neigh = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_policy_deny = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_ha_inactive = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_no_egress_binding = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_build_fail = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_tx_err = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_metadata_err = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_disposition_other = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_enqueue_ok = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_enqueue_inplace = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_enqueue_direct = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_enqueue_copy = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_from_trust = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_from_wan = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_fwd_trust_to_wan = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_fwd_wan_to_trust = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_nat_snat = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_nat_dnat = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_nat_none = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_frame_build_none = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_tcp_rst = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_tx_tcp_rst = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_tcp_fin = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_tcp_synack = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_tcp_zero_window = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_fwd_tcp_fin = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_fwd_tcp_rst = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_fwd_tcp_zero_window = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_bytes_total = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_tx_bytes_total = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_oversized = 0u64;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_rx_max_frame = 0u32;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_tx_max_frame = 0u32;
+    #[cfg(feature = "debug-log")]
+    let mut dbg_seg_needed_but_none = 0u64;
+    let mut prev_rx_total = 0u64;
+    let mut prev_fwd_total = 0u64;
+    let mut stall_prev_fwd = 0u64;
+    let mut stall_reported = false;
+    const DBG_REPORT_INTERVAL_NS: u64 = 1_000_000_000; // 1 second
+    // Throttle for BPF conntrack last_seen refresh (~10s).
+    // Keeps `show security flow session` idle times accurate without
+    // per-second syscall overhead per session.  See issue #333.
+    const CT_REFRESH_INTERVAL_NS: u64 = 10_000_000_000;
+    // Cache BPF map FDs — they don't change during the worker's lifetime.
+    let session_map_fd = bindings
+        .first()
+        .map(|binding| binding.session_map_fd)
+        .unwrap_or(-1);
+    let conntrack_v4_fd = bindings
+        .first()
+        .map(|binding| binding.conntrack_v4_fd)
+        .unwrap_or(-1);
+    let conntrack_v6_fd = bindings
+        .first()
+        .map(|binding| binding.conntrack_v6_fd)
+        .unwrap_or(-1);
+    let mut last_ct_refresh_ns: u64 = 0;
+    cos_status.store(Arc::new(build_worker_cos_statuses(
+        &bindings,
+        forwarding.as_ref(),
+    )));
+    let mut last_cos_status_ns = monotonic_nanos();
+    // #869: worker-runtime telemetry.  Local counters, published to
+    // `runtime_atomics` on the ~1s cadence below.
+    use super::worker_runtime::{
+        WorkerRuntimeCounters, WorkerRuntimeState, current_tid, sample_thread_cpu_ns,
+    };
+    let mut wr_counters = WorkerRuntimeCounters::default();
+    let mut wr_state = WorkerRuntimeState::IdleBlock;
+    let mut wr_last_loop_ns = monotonic_nanos();
+    let mut wr_last_publish_ns = wr_last_loop_ns;
+    const WR_PUBLISH_INTERVAL_NS: u64 = 1_000_000_000;
+    runtime_atomics.set_tid(current_tid());
+    while !stop.load(Ordering::Relaxed) {
+        let loop_now_ns = monotonic_nanos();
+        // #869: attribute elapsed delta to the previous loop's state.
+        {
+            let delta = loop_now_ns.saturating_sub(wr_last_loop_ns);
+            wr_counters.wall_ns = wr_counters.wall_ns.wrapping_add(delta);
+            match wr_state {
+                WorkerRuntimeState::Active => {
+                    wr_counters.active_ns = wr_counters.active_ns.wrapping_add(delta);
+                }
+                WorkerRuntimeState::IdleSpin => {
+                    wr_counters.idle_spin_ns = wr_counters.idle_spin_ns.wrapping_add(delta);
+                }
+                WorkerRuntimeState::IdleBlock => {
+                    wr_counters.idle_block_ns = wr_counters.idle_block_ns.wrapping_add(delta);
+                }
+            }
+            wr_last_loop_ns = loop_now_ns;
+            if loop_now_ns.saturating_sub(wr_last_publish_ns) >= WR_PUBLISH_INTERVAL_NS {
+                // Skip on transient clock_gettime failure (sample == 0):
+                // overwriting a previously-published nonzero value with 0
+                // would make the Prometheus counter go backwards and
+                // break `rate()` queries.
+                let sampled_cpu_ns = sample_thread_cpu_ns();
+                if sampled_cpu_ns != 0 {
+                    wr_counters.thread_cpu_ns = sampled_cpu_ns;
+                }
+                runtime_atomics.publish(&wr_counters);
+                wr_last_publish_ns = loop_now_ns;
+            }
+        }
+        let loop_now_secs = loop_now_ns / 1_000_000_000;
+        let live_validation = shared_validation.load();
+        if **live_validation != validation {
+            validation = **live_validation;
+        }
+        let live_forwarding = shared_forwarding.load_full();
+        let mut rebuild_cos_fast_interfaces = false;
+        if !Arc::ptr_eq(&forwarding, &live_forwarding) {
+            let cos_changed =
+                cos_runtime_config_changed(forwarding.as_ref(), live_forwarding.as_ref());
+            forwarding = live_forwarding;
+            screen_state.update_profiles(forwarding.screen_profiles.clone());
+            sessions.set_timeouts(forwarding.session_timeouts);
+            if cos_changed {
+                reset_worker_cos_runtimes(&mut bindings);
+                rebuild_cos_fast_interfaces = true;
+            }
+        }
+        let live_cos_owner_worker_by_queue = shared_cos_owner_worker_by_queue.load_full();
+        if !Arc::ptr_eq(&cos_owner_worker_by_queue, &live_cos_owner_worker_by_queue) {
+            cos_owner_worker_by_queue = live_cos_owner_worker_by_queue;
+            rebuild_cos_fast_interfaces = true;
+        }
+        let live_cos_owner_live_by_queue = shared_cos_owner_live_by_queue.load_full();
+        if !Arc::ptr_eq(&cos_owner_live_by_queue, &live_cos_owner_live_by_queue) {
+            cos_owner_live_by_queue = live_cos_owner_live_by_queue;
+            rebuild_cos_fast_interfaces = true;
+        }
+        let live_cos_shared_root_leases = shared_cos_root_leases.load_full();
+        if !Arc::ptr_eq(&cos_shared_root_leases, &live_cos_shared_root_leases) {
+            for binding in bindings.iter_mut() {
+                release_all_cos_root_leases(binding);
+                release_all_cos_queue_leases(binding);
+            }
+            cos_shared_root_leases = live_cos_shared_root_leases;
+            rebuild_cos_fast_interfaces = true;
+        }
+        let live_cos_shared_queue_leases = shared_cos_queue_leases.load_full();
+        if !Arc::ptr_eq(&cos_shared_queue_leases, &live_cos_shared_queue_leases) {
+            for binding in bindings.iter_mut() {
+                release_all_cos_queue_leases(binding);
+            }
+            cos_shared_queue_leases = live_cos_shared_queue_leases;
+            rebuild_cos_fast_interfaces = true;
+        }
+        let live_cos_shared_queue_vtime_floors = shared_cos_queue_vtime_floors.load_full();
+        if !Arc::ptr_eq(
+            &cos_shared_queue_vtime_floors,
+            &live_cos_shared_queue_vtime_floors,
+        ) {
+            // #917: Arc-replacement of the V_min floors map.
+            // Each shared_exact queue's per-worker slots default
+            // to NOT_PARTICIPATING in the new Arc. Workers will
+            // re-publish their committed vtime on the next
+            // commit-boundary publish; until then peers reading
+            // this slot see "not participating" and skip it in
+            // V_min reduction (per plan §3.4 / §3.7 lifecycle
+            // rules).
+            cos_shared_queue_vtime_floors = live_cos_shared_queue_vtime_floors;
+            rebuild_cos_fast_interfaces = true;
+        }
+        if rebuild_cos_fast_interfaces {
+            let cos_owner_live_by_tx_ifindex = build_worker_cos_owner_live_by_tx_ifindex(
+                bindings
+                    .iter()
+                    .map(|binding| (binding.ifindex, binding.live.clone())),
+            );
+            let cos_fast_interfaces = build_worker_cos_fast_interfaces(
+                forwarding.as_ref(),
+                worker_id,
+                &cos_owner_live_by_tx_ifindex,
+                cos_owner_worker_by_queue.as_ref(),
+                cos_owner_live_by_queue.as_ref(),
+                cos_shared_root_leases.as_ref(),
+                cos_shared_queue_leases.as_ref(),
+                cos_shared_queue_vtime_floors.as_ref(),
+            );
+            for binding in bindings.iter_mut() {
+                binding.cos_fast_interfaces = cos_fast_interfaces.clone();
+            }
+        }
+        let ha_runtime = ha_state.load();
+        // Only apply commands when pending — avoids lock overhead on
+        // every loop iteration in the common (empty-queue) case.
+        let has_commands = commands.try_lock().map(|q| !q.is_empty()).unwrap_or(false);
+        let command_results = if has_commands {
+            apply_worker_commands(
+                &commands,
+                &mut sessions,
+                session_map_fd,
+                conntrack_v4_fd,
+                conntrack_v6_fd,
+                &forwarding,
+                ha_runtime.as_ref(),
+                &dynamic_neighbors,
+            )
+        } else {
+            WorkerCommandResults {
+                cancelled_keys: Vec::new(),
+                exported_sequences: Vec::new(),
+                shaped_tx_requests: Vec::new(),
+                vacate_all_shared_exact_slots: false,
+            }
+        };
+        let WorkerCommandResults {
+            cancelled_keys,
+            exported_sequences,
+            shaped_tx_requests,
+            vacate_all_shared_exact_slots,
+        } = command_results;
+        // #941 Work item C: HA-demotion vacate. The
+        // VacateAllSharedExactSlots WorkerCommand cannot be processed
+        // inside `apply_worker_commands` (no BindingWorker access);
+        // it sets this flag and the dispatch happens here, where we
+        // hold `&mut bindings`. Single-writer invariant: only this
+        // worker writes its own slots.
+        if vacate_all_shared_exact_slots {
+            for binding in bindings.iter_mut() {
+                vacate_all_shared_exact_slots_for_binding(binding);
+            }
+        }
+        if !shaped_tx_requests.is_empty() {
+            apply_worker_shaped_tx_requests(
+                &mut bindings,
+                forwarding.as_ref(),
+                &binding_lookup,
+                loop_now_ns,
+                shaped_tx_requests,
+            );
+        }
+        if !cancelled_keys.is_empty() {
+            for key in &cancelled_keys {
+                for binding in bindings.iter_mut() {
+                    cancel_queued_flow_on_binding(binding, key, key);
+                }
+                if let Some((decision, metadata, origin)) = sessions.entry_with_origin(key) {
+                    // Demotion keeps the session in the standby table, but the
+                    // stale owner must stop advertising local XSK redirect
+                    // aliases immediately or XDP will keep steering packets to
+                    // the old node after RG handoff.
+                    delete_session_map_redirect_for_session(
+                        session_map_fd,
+                        key,
+                        decision,
+                        &metadata,
+                        origin,
+                    );
+                }
+            }
+        }
+        heartbeat.store(loop_now_ns, Ordering::Relaxed);
+        let expired_entries = sessions.expire_stale_entries(loop_now_ns);
+        let expired = expired_entries.len() as u64;
+        for expired_entry in expired_entries {
+            delete_session_map_entry_for_removed_session_with_origin(
+                session_map_fd,
+                &expired_entry.key,
+                expired_entry.decision,
+                &expired_entry.metadata,
+                expired_entry.origin,
+                conntrack_v4_fd,
+                conntrack_v6_fd,
+            );
+        }
+        if expired > 0 {
+            if let Some(binding) = bindings.first() {
+                binding
+                    .live
+                    .session_expires
+                    .fetch_add(expired, Ordering::Relaxed);
+            }
+        }
+        // Periodically refresh last_seen in BPF conntrack entries so Go-side
+        // callers of IterateSessions (CLI, gRPC, Prometheus) see accurate
+        // session idle times.  Issue #333.
+        if loop_now_ns.saturating_sub(last_ct_refresh_ns) >= CT_REFRESH_INTERVAL_NS {
+            last_ct_refresh_ns = loop_now_ns;
+            refresh_bpf_conntrack_last_seen(
+                conntrack_v4_fd,
+                conntrack_v6_fd,
+                &sessions,
+                loop_now_ns,
+            );
+        }
+        // Check if fabric links were updated by the coordinator (e.g. after
+        // RG failover when peer MAC was resolved). If so, rebuild the
+        // forwarding Arc with the new fabric links so fabric redirect works.
+        {
+            let live_fabrics = shared_fabrics.load();
+            if !live_fabrics.is_empty() && live_fabrics.as_ref() != &forwarding.fabrics {
+                let mut updated = (*forwarding).clone();
+                updated.fabrics = live_fabrics.as_ref().clone();
+                forwarding = Arc::new(updated);
+            }
+        }
+        let mut did_work = false;
+        let mut dbg_poll = DebugPollCounters::default();
+        for offset in 0..bindings.len() {
+            let idx = if bindings.is_empty() {
+                0
+            } else {
+                (poll_start + offset) % bindings.len()
+            };
+            if poll_binding(
+                idx,
+                &mut bindings,
+                &binding_lookup,
+                &mut sessions,
+                &mut screen_state,
+                validation,
+                loop_now_ns,
+                loop_now_secs,
+                ha_startup_grace_until_secs,
+                &forwarding,
+                ha_runtime.as_ref(),
+                &dynamic_neighbors,
+                &shared_sessions,
+                &shared_nat_sessions,
+                &shared_forward_wire_sessions,
+                &shared_owner_rg_indexes,
+                slow_path.as_ref(),
+                &local_tunnel_deliveries,
+                &recent_exceptions,
+                &recent_session_deltas,
+                &last_resolution,
+                &peer_worker_commands,
+                worker_id,
+                worker_commands_by_id.as_ref(),
+                &mut shared_recycles,
+                &dnat_fds,
+                conntrack_v4_fd,
+                conntrack_v6_fd,
+                &mut dbg_poll,
+                &rg_epochs,
+                cos_owner_worker_by_queue.as_ref(),
+                cos_owner_live_by_queue.as_ref(),
+            ) {
+                did_work = true;
+            }
+        }
+        crate::filter::flush_recorded_filter_counters();
+        dbg_rx_total += dbg_poll.rx;
+        #[cfg(feature = "debug-log")]
+        {
+            dbg_tx_total += dbg_poll.tx;
+        }
+        dbg_forward_total += dbg_poll.forward;
+        #[cfg(feature = "debug-log")]
+        {
+            dbg_local_total += dbg_poll.local;
+            dbg_session_hit += dbg_poll.session_hit;
+            dbg_session_miss += dbg_poll.session_miss;
+            dbg_session_create += dbg_poll.session_create;
+            dbg_no_route += dbg_poll.no_route;
+            dbg_missing_neigh += dbg_poll.missing_neigh;
+            dbg_policy_deny += dbg_poll.policy_deny;
+            dbg_ha_inactive += dbg_poll.ha_inactive;
+            dbg_no_egress_binding += dbg_poll.no_egress_binding;
+            dbg_build_fail += dbg_poll.build_fail;
+            dbg_tx_err += dbg_poll.tx_err;
+            dbg_metadata_err += dbg_poll.metadata_err;
+        }
+        #[cfg(feature = "debug-log")]
+        {
+            dbg_disposition_other += dbg_poll.disposition_other;
+            dbg_enqueue_ok += dbg_poll.enqueue_ok;
+            dbg_enqueue_inplace += dbg_poll.enqueue_inplace;
+            dbg_enqueue_direct += dbg_poll.enqueue_direct;
+            dbg_enqueue_copy += dbg_poll.enqueue_copy;
+            dbg_rx_from_trust += dbg_poll.rx_from_trust;
+            dbg_rx_from_wan += dbg_poll.rx_from_wan;
+            dbg_fwd_trust_to_wan += dbg_poll.fwd_trust_to_wan;
+            dbg_fwd_wan_to_trust += dbg_poll.fwd_wan_to_trust;
+            dbg_nat_snat += dbg_poll.nat_applied_snat;
+            dbg_nat_dnat += dbg_poll.nat_applied_dnat;
+            dbg_nat_none += dbg_poll.nat_applied_none;
+            dbg_frame_build_none += dbg_poll.frame_build_none;
+        }
+        #[cfg(feature = "debug-log")]
+        {
+            dbg_rx_tcp_rst += dbg_poll.rx_tcp_rst;
+            dbg_rx_tcp_fin += dbg_poll.rx_tcp_fin;
+            dbg_rx_tcp_synack += dbg_poll.rx_tcp_synack;
+            dbg_rx_tcp_zero_window += dbg_poll.rx_tcp_zero_window;
+            dbg_fwd_tcp_fin += dbg_poll.fwd_tcp_fin;
+            dbg_fwd_tcp_rst += dbg_poll.fwd_tcp_rst;
+            dbg_fwd_tcp_zero_window += dbg_poll.fwd_tcp_zero_window;
+        }
+        #[cfg(feature = "debug-log")]
+        {
+            dbg_rx_bytes_total += dbg_poll.rx_bytes_total;
+            dbg_tx_bytes_total += dbg_poll.tx_bytes_total;
+            dbg_rx_oversized += dbg_poll.rx_oversized;
+            if dbg_poll.rx_max_frame > dbg_rx_max_frame {
+                dbg_rx_max_frame = dbg_poll.rx_max_frame;
+            }
+            if dbg_poll.tx_max_frame > dbg_tx_max_frame {
+                dbg_tx_max_frame = dbg_poll.tx_max_frame;
+            }
+            dbg_seg_needed_but_none += dbg_poll.seg_needed_but_none;
+        }
+        if !bindings.is_empty() {
+            poll_start = (poll_start + 1) % bindings.len();
+        }
+        if loop_now_ns.saturating_sub(last_cos_status_ns) >= COS_STATUS_INTERVAL_NS {
+            cos_status.store(Arc::new(build_worker_cos_statuses(
+                &bindings,
+                forwarding.as_ref(),
+            )));
+            last_cos_status_ns = loop_now_ns;
+        }
+        if !exported_sequences.is_empty() {
+            while sessions.has_pending_deltas() {
+                let deltas = sessions.drain_deltas(256);
+                purge_queued_flows_for_closed_deltas(&mut bindings, &deltas);
+                if let Some(binding) = bindings.first() {
+                    let ident = binding.identity();
+                    flush_session_deltas(
+                        &ident,
+                        &binding.live,
+                        binding.session_map_fd,
+                        conntrack_v4_fd,
+                        conntrack_v6_fd,
+                        &deltas,
+                        &shared_sessions,
+                        &shared_nat_sessions,
+                        &shared_forward_wire_sessions,
+                        &shared_owner_rg_indexes,
+                        &recent_session_deltas,
+                        &peer_worker_commands,
+                        &event_stream,
+                        forwarding.as_ref(),
+                    );
+                }
+            }
+            if let Some(sequence) = exported_sequences.iter().copied().max() {
+                session_export_ack.store(sequence, Ordering::Release);
+            }
+        } else if sessions.has_pending_deltas() {
+            let deltas = sessions.drain_deltas(256);
+            purge_queued_flows_for_closed_deltas(&mut bindings, &deltas);
+            if let Some(binding) = bindings.first() {
+                let ident = binding.identity();
+                flush_session_deltas(
+                    &ident,
+                    &binding.live,
+                    binding.session_map_fd,
+                    conntrack_v4_fd,
+                    conntrack_v6_fd,
+                    &deltas,
+                    &shared_sessions,
+                    &shared_nat_sessions,
+                    &shared_forward_wire_sessions,
+                    &shared_owner_rg_indexes,
+                    &recent_session_deltas,
+                    &peer_worker_commands,
+                    &event_stream,
+                    forwarding.as_ref(),
+                );
+            }
+        }
+        // Debug: periodic summary report
+        {
+            let elapsed = loop_now_ns.saturating_sub(dbg_last_report_ns);
+            if elapsed >= DBG_REPORT_INTERVAL_NS {
+                #[cfg(feature = "debug-log")]
+                let secs = elapsed as f64 / 1_000_000_000.0;
+                let session_count = sessions.len();
+                let mut binding_summary = String::new();
+                for (i, b) in bindings.iter().enumerate() {
+                    use std::fmt::Write;
+                    let fill_pending = b.device.pending();
+                    let rx_avail = b.rx.available_relaxed();
+                    let xsk_stats = b.device.statistics_v2().ok();
+                    let inflight_recycles = b.in_flight_prepared_recycles.len() as u32;
+                    let scratch_recycle_len = b.scratch_recycle.len() as u32;
+                    let ptx_prepared = b.pending_tx_prepared.len() as u32;
+                    let ptx_local = b.pending_tx_local.len() as u32;
+                    let total_accounted = b.pending_fill_frames.len() as u32
+                        + fill_pending
+                        + rx_avail
+                        + b.free_tx_frames.len() as u32
+                        + b.outstanding_tx
+                        + inflight_recycles
+                        + scratch_recycle_len
+                        + ptx_prepared; // prepared TX holds UMEM frames
+                    let expected_total = b.umem.total_frames();
+                    let _ = write!(
+                        binding_summary,
+                        " [{}:if{}q{} pfill={} fring={} rxring={} free_tx={} otx={} ifl={} scr={} ptxp={} ptxl={} total={}/{} fill_ok={} polls={} bp={} rx_empty={} wake={}",
+                        i,
+                        b.ifindex,
+                        b.queue_id,
+                        b.pending_fill_frames.len(),
+                        fill_pending,
+                        rx_avail,
+                        b.free_tx_frames.len(),
+                        b.outstanding_tx,
+                        inflight_recycles,
+                        scratch_recycle_len,
+                        ptx_prepared,
+                        ptx_local,
+                        total_accounted,
+                        expected_total,
+                        b.dbg_fill_submitted,
+                        b.dbg_poll_cycles,
+                        b.dbg_backpressure,
+                        b.dbg_rx_empty,
+                        b.dbg_rx_wakeups,
+                    );
+                    // TX pipeline debug counters
+                    #[cfg(feature = "debug-log")]
+                    {
+                        dbg_tx_tcp_rst += b.dbg_tx_tcp_rst;
+                    }
+                    let _ = write!(
+                        binding_summary,
+                        " TX:ring_sub={}/ring_full={}/compl={}/sendto={}/err={}/eagain={}/enobufs={}/bp_overflow={}/cos_overflow={}",
+                        b.dbg_tx_ring_submitted,
+                        b.dbg_tx_ring_full,
+                        b.dbg_completions_reaped,
+                        b.dbg_sendto_calls,
+                        b.dbg_sendto_err,
+                        b.dbg_sendto_eagain,
+                        b.dbg_sendto_enobufs,
+                        b.dbg_bound_pending_overflow,
+                        b.dbg_cos_queue_overflow,
+                    );
+                    #[cfg(feature = "debug-log")]
+                    let _ = write!(binding_summary, "/rst={}", b.dbg_tx_tcp_rst);
+                    if let Some(s) = xsk_stats {
+                        let _ = write!(
+                            binding_summary,
+                            " xsk:drop={}/inv={}/rfull={}/fempty={}/tinv={}/tempty={}",
+                            s.rx_dropped,
+                            s.rx_invalid_descs,
+                            s.rx_ring_full,
+                            s.rx_fill_ring_empty_descs,
+                            s.tx_invalid_descs,
+                            s.tx_ring_empty_descs,
+                        );
+                    }
+                    // Socket error check (SO_ERROR) — detect kernel-side errors
+                    {
+                        let fd = b.rx.as_raw_fd();
+                        let mut so_err: c_int = 0;
+                        let mut so_err_len: libc::socklen_t = core::mem::size_of::<c_int>() as _;
+                        let rc = unsafe {
+                            libc::getsockopt(
+                                fd,
+                                libc::SOL_SOCKET,
+                                libc::SO_ERROR,
+                                &mut so_err as *mut c_int as *mut c_void,
+                                &mut so_err_len,
+                            )
+                        };
+                        if rc == 0 && so_err != 0 {
+                            let _ = write!(binding_summary, " SO_ERR={so_err}");
+                        }
+                    }
+                    // Ring diagnostics from xsk_ffi API
+                    if cfg!(feature = "debug-log") {
+                        let _ = write!(
+                            binding_summary,
+                            " RING:rx_nz={}/rx_max={}/fill_pend={}/dev_avail={} RX_WAKE:ok={}/err={}/errno={}",
+                            b.dbg_rx_avail_nonzero,
+                            b.dbg_rx_avail_max,
+                            b.dbg_fill_pending,
+                            b.dbg_device_avail,
+                            b.dbg_rx_wake_sendto_ok,
+                            b.dbg_rx_wake_sendto_err,
+                            b.dbg_rx_wake_sendto_errno,
+                        );
+                        // Direct mmap diagnosis: read raw ring producer/consumer
+                        if let Some((rxp, rxc, frp, frc, txp, txc, crp, crc)) =
+                            diagnose_raw_ring_state(b.rx.as_raw_fd())
+                        {
+                            let _ = write!(
+                                binding_summary,
+                                " RAW:rxP={rxp}/rxC={rxc}/frP={frp}/frC={frc}/txP={txp}/txC={txc}/crP={crp}/crC={crc}"
+                            );
+                        }
+                    }
+                    // Frame leak detection
+                    if total_accounted != expected_total {
+                        let _ = write!(
+                            binding_summary,
+                            " FRAME_LEAK:{}",
+                            expected_total as i64 - total_accounted as i64,
+                        );
+                    }
+                    binding_summary.push(']');
+                }
+                #[cfg(feature = "debug-log")]
+                eprintln!(
+                    "DBG w{}: {:.1}s rx={} tx={} fwd={} local={} sess_hit={} sess_miss={} sess_create={} \
+                     no_route={} miss_neigh={} pol_deny={} ha_inact={} no_egress={} build_fail={} \
+                     tx_err={} meta_err={} other={} enq_ok={} enq_ip={} enq_dir={} enq_cp={} sessions={} \
+                     DIR:trust_rx={}/wan_rx={}/t2w={}/w2t={} NAT:snat={}/dnat={}/none={}/bld_none={} RST:rx={}/tx={} \
+                     SIZE:rx_avg={}/rx_max={}/tx_avg={}/tx_max={}/rx_over={}/seg_miss={} \
+                     TCP_RX:fin={}/synack={}/zwin={} TCP_FWD:fin={}/rst={}/zwin={} \
+                     CSUM:verified={}/bad_ip={}/bad_l4={} \
+                     SESS_BPF:verify_ok={}/verify_fail={}/bpf_entries={} bindings:{}",
+                    worker_id,
+                    secs,
+                    dbg_rx_total,
+                    dbg_tx_total,
+                    dbg_forward_total,
+                    dbg_local_total,
+                    dbg_session_hit,
+                    dbg_session_miss,
+                    dbg_session_create,
+                    dbg_no_route,
+                    dbg_missing_neigh,
+                    dbg_policy_deny,
+                    dbg_ha_inactive,
+                    dbg_no_egress_binding,
+                    dbg_build_fail,
+                    dbg_tx_err,
+                    dbg_metadata_err,
+                    dbg_disposition_other,
+                    dbg_enqueue_ok,
+                    dbg_enqueue_inplace,
+                    dbg_enqueue_direct,
+                    dbg_enqueue_copy,
+                    session_count,
+                    dbg_rx_from_trust,
+                    dbg_rx_from_wan,
+                    dbg_fwd_trust_to_wan,
+                    dbg_fwd_wan_to_trust,
+                    dbg_nat_snat,
+                    dbg_nat_dnat,
+                    dbg_nat_none,
+                    dbg_frame_build_none,
+                    dbg_rx_tcp_rst,
+                    dbg_tx_tcp_rst,
+                    if dbg_rx_total > 0 {
+                        dbg_rx_bytes_total / dbg_rx_total
+                    } else {
+                        0
+                    },
+                    dbg_rx_max_frame,
+                    if dbg_enqueue_ok > 0 {
+                        dbg_tx_bytes_total / dbg_enqueue_ok
+                    } else {
+                        0
+                    },
+                    dbg_tx_max_frame,
+                    dbg_rx_oversized,
+                    dbg_seg_needed_but_none,
+                    dbg_rx_tcp_fin,
+                    dbg_rx_tcp_synack,
+                    dbg_rx_tcp_zero_window,
+                    dbg_fwd_tcp_fin,
+                    dbg_fwd_tcp_rst,
+                    dbg_fwd_tcp_zero_window,
+                    CSUM_VERIFIED_TOTAL.swap(0, Ordering::Relaxed),
+                    CSUM_BAD_IP_TOTAL.swap(0, Ordering::Relaxed),
+                    CSUM_BAD_L4_TOTAL.swap(0, Ordering::Relaxed),
+                    SESSION_PUBLISH_VERIFY_OK.swap(0, Ordering::Relaxed),
+                    SESSION_PUBLISH_VERIFY_FAIL.swap(0, Ordering::Relaxed),
+                    if let Some(b) = bindings.first() {
+                        count_bpf_session_entries(b.session_map_fd)
+                    } else {
+                        0
+                    },
+                    binding_summary,
+                );
+                // Non-debug builds: no per-second stats dump (use debug-log feature for verbose output).
+                // Print XDP shim fallback stats — tells us WHY packets stop
+                // being redirected to XSK.
+                if cfg!(feature = "debug-log") {
+                    if let Some(stats) = read_fallback_stats() {
+                        if !stats.is_empty() {
+                            let s: Vec<String> =
+                                stats.iter().map(|(n, v)| format!("{n}={v}")).collect();
+                            eprintln!("DBG w{}: XDP_FALLBACK: {}", worker_id, s.join(" "));
+                        }
+                    }
+                }
+                // Save prev counters BEFORE reset for stall detection below
+                if cfg!(feature = "debug-log") {
+                    prev_rx_total = dbg_rx_total;
+                    prev_fwd_total = dbg_forward_total;
+                }
+                dbg_last_report_ns = loop_now_ns;
+                dbg_rx_total = 0;
+                #[cfg(feature = "debug-log")]
+                {
+                    dbg_tx_total = 0;
+                }
+                dbg_forward_total = 0;
+                #[cfg(feature = "debug-log")]
+                {
+                    dbg_local_total = 0;
+                    dbg_session_hit = 0;
+                    dbg_session_miss = 0;
+                    dbg_session_create = 0;
+                    dbg_no_route = 0;
+                    dbg_missing_neigh = 0;
+                    dbg_policy_deny = 0;
+                    dbg_ha_inactive = 0;
+                    dbg_no_egress_binding = 0;
+                    dbg_build_fail = 0;
+                    dbg_tx_err = 0;
+                    dbg_metadata_err = 0;
+                }
+                #[cfg(feature = "debug-log")]
+                {
+                    dbg_disposition_other = 0;
+                    dbg_enqueue_ok = 0;
+                    dbg_enqueue_inplace = 0;
+                    dbg_enqueue_direct = 0;
+                    dbg_enqueue_copy = 0;
+                    dbg_rx_from_trust = 0;
+                    dbg_rx_from_wan = 0;
+                    dbg_fwd_trust_to_wan = 0;
+                    dbg_fwd_wan_to_trust = 0;
+                }
+                #[cfg(feature = "debug-log")]
+                {
+                    dbg_rx_bytes_total = 0;
+                    dbg_tx_bytes_total = 0;
+                    dbg_rx_oversized = 0;
+                    dbg_rx_max_frame = 0;
+                    dbg_tx_max_frame = 0;
+                    dbg_seg_needed_but_none = 0;
+                }
+                // Stall detection: stall_prev_fwd is PREVIOUS interval's fwd count,
+                // prev_fwd_total is THIS interval's fwd count (saved before reset).
+                if cfg!(feature = "debug-log") {
+                    if stall_prev_fwd > 10 && prev_fwd_total == 0 && !stall_reported {
+                        stall_reported = true;
+                        eprintln!(
+                            "DBG STALL_DETECTED: w{} two_ago_fwd={} this_interval_fwd={} this_interval_rx={} sessions={}",
+                            worker_id, stall_prev_fwd, prev_fwd_total, prev_rx_total, session_count
+                        );
+                        // Dump comprehensive per-binding state at stall moment
+                        for (si, sb) in bindings.iter().enumerate() {
+                            use std::fmt::Write;
+                            let fill_p = sb.device.pending();
+                            let rx_a = sb.rx.available_relaxed();
+                            let ifl = sb.in_flight_prepared_recycles.len() as u32;
+                            let ptxp = sb.pending_tx_prepared.len() as u32;
+                            let ptxl = sb.pending_tx_local.len() as u32;
+                            let total = sb.pending_fill_frames.len() as u32
+                                + fill_p
+                                + rx_a
+                                + sb.free_tx_frames.len() as u32
+                                + sb.outstanding_tx
+                                + ifl
+                                + sb.scratch_recycle.len() as u32
+                                + ptxp;
+                            let raw = diagnose_raw_ring_state(sb.rx.as_raw_fd());
+                            let mut stall_line = format!(
+                                "DBG STALL_BINDING[{}]: if={} q={} pfill={} fring={} rxring={} free_tx={} otx={} ifl={} ptxp={} ptxl={} total={}/{}",
+                                si,
+                                sb.ifindex,
+                                sb.queue_id,
+                                sb.pending_fill_frames.len(),
+                                fill_p,
+                                rx_a,
+                                sb.free_tx_frames.len(),
+                                sb.outstanding_tx,
+                                ifl,
+                                ptxp,
+                                ptxl,
+                                total,
+                                sb.umem.total_frames(),
+                            );
+                            if let Some((rxp, rxc, frp, frc, txp, txc, crp, crc)) = raw {
+                                let _ = write!(
+                                    stall_line,
+                                    " RAW:rxP={rxp}/rxC={rxc}/frP={frp}/frC={frc}/txP={txp}/txC={txc}/crP={crp}/crC={crc}"
+                                );
+                            }
+                            if let Ok(Some(stats)) = sb.device.statistics_v2().map(Some) {
+                                let _ = write!(
+                                    stall_line,
+                                    " xsk:drop={}/rfull={}/fempty={}/tempty={}",
+                                    stats.rx_dropped,
+                                    stats.rx_ring_full,
+                                    stats.rx_fill_ring_empty_descs,
+                                    stats.tx_ring_empty_descs
+                                );
+                            }
+                            eprintln!("{stall_line}");
+                        }
+                        // Dump all session keys for this worker
+                        let mut sess_dump = String::new();
+                        let mut count = 0;
+                        sessions.iter_with_origin(|key, decision, metadata, origin| {
+                            if count < 20 {
+                                use std::fmt::Write;
+                                let _ = write!(
+                                    sess_dump,
+                                    "\n  SESS: {}:{} -> {}:{} proto={} nat=({:?},{:?}) is_rev={} origin={}",
+                                    key.src_ip,
+                                    key.src_port,
+                                    key.dst_ip,
+                                    key.dst_port,
+                                    key.protocol,
+                                    decision.nat.rewrite_src,
+                                    decision.nat.rewrite_dst,
+                                    metadata.is_reverse,
+                                    origin.as_str(),
+                                );
+                                count += 1;
+                            }
+                        });
+                        if !sess_dump.is_empty() {
+                            eprintln!("DBG STALL_SESSIONS:{sess_dump}");
+                        }
+                        // Dump fallback stats at stall time
+                        if let Some(stats) = read_fallback_stats() {
+                            if !stats.is_empty() {
+                                let s: Vec<String> =
+                                    stats.iter().map(|(n, v)| format!("{n}={v}")).collect();
+                                eprintln!("DBG STALL_FALLBACK: {}", s.join(" "));
+                            }
+                        }
+                        // Also dump BPF session count
+                        if let Some(b) = bindings.first() {
+                            eprintln!(
+                                "DBG STALL_BPF_SESSIONS: entries={}",
+                                count_bpf_session_entries(b.session_map_fd)
+                            );
+                        }
+                    } else if prev_fwd_total > 0 {
+                        stall_reported = false;
+                    }
+                    stall_prev_fwd = prev_fwd_total;
+                }
+                #[cfg(feature = "debug-log")]
+                {
+                    dbg_nat_snat = 0;
+                    dbg_nat_dnat = 0;
+                    dbg_nat_none = 0;
+                    dbg_frame_build_none = 0;
+                }
+                #[cfg(feature = "debug-log")]
+                {
+                    dbg_rx_tcp_rst = 0;
+                    dbg_tx_tcp_rst = 0;
+                    dbg_rx_tcp_fin = 0;
+                    dbg_rx_tcp_synack = 0;
+                    dbg_rx_tcp_zero_window = 0;
+                    dbg_fwd_tcp_fin = 0;
+                    dbg_fwd_tcp_rst = 0;
+                    dbg_fwd_tcp_zero_window = 0;
+                }
+                for b in bindings.iter_mut() {
+                    // #802: publish ring-pressure counters into BindingLiveState
+                    // BEFORE resetting the worker-local window. The worker-local
+                    // counters (b.dbg_tx_ring_full, etc.) are accumulated by the
+                    // hot path and reset each ~1s debug tick; without this
+                    // publish they'd never be visible outside the worker thread.
+                    // fetch_add is used because the atomic holds the cumulative
+                    // total while the local counter holds only the current
+                    // window. Relaxed is sufficient — diagnostic counters, no
+                    // synchronization contract.
+                    if b.dbg_tx_ring_full != 0 {
+                        b.live
+                            .dbg_tx_ring_full
+                            .fetch_add(b.dbg_tx_ring_full, Ordering::Relaxed);
+                    }
+                    if b.dbg_sendto_enobufs != 0 {
+                        b.live
+                            .dbg_sendto_enobufs
+                            .fetch_add(b.dbg_sendto_enobufs, Ordering::Relaxed);
+                    }
+                    if b.dbg_bound_pending_overflow != 0 {
+                        b.live
+                            .dbg_bound_pending_overflow
+                            .fetch_add(b.dbg_bound_pending_overflow, Ordering::Relaxed);
+                    }
+                    if b.dbg_cos_queue_overflow != 0 {
+                        b.live
+                            .dbg_cos_queue_overflow
+                            .fetch_add(b.dbg_cos_queue_overflow, Ordering::Relaxed);
+                    }
+                    // #802: kernel xdp_statistics.rx_fill_ring_empty_descs is
+                    // already absolute (kernel-cumulative), so publish with
+                    // store() not fetch_add. Sampling failures are silently
+                    // ignored — the atomic simply retains its last good value.
+                    if let Ok(stats) = b.device.statistics_v2() {
+                        b.live
+                            .rx_fill_ring_empty_descs
+                            .store(stats.rx_fill_ring_empty_descs, Ordering::Relaxed);
+                    }
+                    // #802: outstanding_tx is a transient gauge on BindingWorker
+                    // (current in-flight TX). Publish to the existing atomic
+                    // mirror on BindingLiveState so the snapshot reader sees
+                    // a recent value. store() because it's a gauge, not a
+                    // counter.
+                    b.live
+                        .debug_outstanding_tx
+                        .store(b.outstanding_tx, Ordering::Relaxed);
+                    // #878: publish UMEM in-flight gauge as a single atomic
+                    // so the daemon's `show chassis forwarding` Buffer% can
+                    // divide by `umem_total_frames` without torn-load risk.
+                    // Computed in this thread from worker-local state, so
+                    // the inputs are mutually consistent at sample time.
+                    //
+                    // "Idle" frames are: free_tx_frames (worker's TX-available
+                    // pool), pending_fill_frames (worker's queue waiting to
+                    // push to the kernel's fill ring), AND fill_pending (the
+                    // kernel's fill ring itself, which holds frames the
+                    // kernel can place RX data into — those are NOT in
+                    // flight). Without subtracting fill_pending the gauge
+                    // reads ~70-80% at idle because AF_XDP keeps the fill
+                    // ring pre-populated by design.
+                    let total = b.umem.total_frames();
+                    let free_tx = b.free_tx_frames.len() as u32;
+                    let pending_fill = b.pending_fill_frames.len() as u32;
+                    let kernel_fill = b.device.pending();
+                    let inflight = total
+                        .saturating_sub(free_tx)
+                        .saturating_sub(pending_fill)
+                        .saturating_sub(kernel_fill);
+                    b.live
+                        .umem_inflight_frames
+                        .store(inflight, Ordering::Relaxed);
+
+                    b.dbg_fill_submitted = 0;
+                    b.dbg_fill_failed = 0;
+                    b.dbg_poll_cycles = 0;
+                    b.dbg_backpressure = 0;
+                    b.dbg_rx_empty = 0;
+                    b.dbg_rx_wakeups = 0;
+                    b.dbg_tx_ring_submitted = 0;
+                    b.dbg_tx_ring_full = 0;
+                    b.dbg_completions_reaped = 0;
+                    b.dbg_sendto_calls = 0;
+                    b.dbg_sendto_err = 0;
+                    b.dbg_sendto_eagain = 0;
+                    b.dbg_sendto_enobufs = 0;
+                    b.dbg_bound_pending_overflow = 0;
+                    b.dbg_cos_queue_overflow = 0;
+                    #[cfg(feature = "debug-log")]
+                    {
+                        b.dbg_tx_tcp_rst = 0;
+                    }
+                    b.dbg_rx_avail_nonzero = 0;
+                    b.dbg_rx_avail_max = 0;
+                    b.dbg_rx_wake_sendto_ok = 0;
+                    b.dbg_rx_wake_sendto_err = 0;
+                    b.dbg_rx_wake_sendto_errno = 0;
+                }
+            }
+        }
+        if did_work {
+            idle_iters = 0;
+            // #869: classify this iteration for next-loop-top accounting.
+            wr_state = WorkerRuntimeState::Active;
+            wr_counters.work_loops = wr_counters.work_loops.wrapping_add(1);
+            continue;
+        }
+        idle_iters = idle_iters.saturating_add(1);
+        wr_counters.idle_loops = wr_counters.idle_loops.wrapping_add(1);
+        match poll_mode {
+            crate::PollMode::BusyPoll => {
+                if idle_iters <= IDLE_SPIN_ITERS {
+                    wr_state = WorkerRuntimeState::IdleSpin;
+                    std::hint::spin_loop();
+                } else {
+                    wr_state = WorkerRuntimeState::IdleBlock;
+                    thread::sleep(Duration::from_micros(IDLE_SLEEP_US));
+                }
+            }
+            crate::PollMode::Interrupt => {
+                // Interrupt mode still needs a short local spin before blocking.
+                // Firewall-local TCP flows are ACK-latency-sensitive; blocking
+                // immediately on the first empty poll collapses cwnd badly.
+                if idle_iters <= IDLE_SPIN_ITERS {
+                    wr_state = WorkerRuntimeState::IdleSpin;
+                    std::hint::spin_loop();
+                } else if !interrupt_poll_fds.is_empty() {
+                    wr_state = WorkerRuntimeState::IdleBlock;
+                    for pfd in &mut interrupt_poll_fds {
+                        pfd.revents = 0;
+                    }
+                    unsafe {
+                        libc::poll(
+                            interrupt_poll_fds.as_mut_ptr(),
+                            interrupt_poll_fds.len() as libc::nfds_t,
+                            INTERRUPT_POLL_TIMEOUT_MS,
+                        );
+                    }
+                } else {
+                    wr_state = WorkerRuntimeState::IdleBlock;
+                    thread::sleep(Duration::from_millis(INTERRUPT_POLL_TIMEOUT_MS as u64));
+                }
+            }
+        }
+    }
+    crate::filter::flush_recorded_filter_counters();
+    for binding in bindings.iter_mut() {
+        release_all_cos_root_leases(binding);
+        release_all_cos_queue_leases(binding);
+    }
+    cos_status.store(Arc::new(build_worker_cos_statuses(
+        &bindings,
+        forwarding.as_ref(),
+    )));
+    heartbeat.store(monotonic_nanos(), Ordering::Relaxed);
+}
+
+fn apply_worker_shaped_tx_requests(
+    bindings: &mut [BindingWorker],
+    forwarding: &ForwardingState,
+    binding_lookup: &WorkerBindingLookup,
+    now_ns: u64,
+    requests: Vec<TxRequest>,
+) {
+    for req in requests {
+        let binding_index = bindings
+            .first()
+            .and_then(|binding| binding.cos_fast_interfaces.get(&req.egress_ifindex))
+            .and_then(|iface_fast| {
+                binding_lookup
+                    .first_by_if
+                    .get(&iface_fast.tx_ifindex)
+                    .copied()
+            })
+            .or_else(|| {
+                let tx_ifindex = resolve_tx_binding_ifindex(forwarding, req.egress_ifindex);
+                binding_lookup.first_by_if.get(&tx_ifindex).copied()
+            });
+        let Some(binding) = binding_index.and_then(|idx| bindings.get_mut(idx)) else {
+            if let Some(binding) = bindings.first_mut() {
+                binding.live.tx_errors.fetch_add(1, Ordering::Relaxed);
+                // #710: dedicated counter — a cross-worker shaped TX
+                // request arrived for an egress this worker has no
+                // binding to drain. Subset of tx_errors.
+                binding
+                    .live
+                    .no_owner_binding_drops
+                    .fetch_add(1, Ordering::Relaxed);
+            }
+            if cfg!(feature = "debug-log") {
+                debug_log!(
+                    "DBG COS_OWNER_MISSING_BINDING: egress_ifindex={}",
+                    req.egress_ifindex,
+                );
+            }
+            continue;
+        };
+        match enqueue_local_into_cos(binding, forwarding, req, now_ns) {
+            Ok(()) => {}
+            Err(req) => {
+                binding.pending_tx_local.push_back(req);
+                bound_pending_tx_local(binding);
+            }
+        }
+    }
+}
+
+
+// Empirical per-worker sustained exact throughput ceiling in bytes/sec. A
+// single owner worker can reliably drive an exact queue up to about this rate
+// before the drain loop backs up and throughput collapses (the collapse case
+// that motivated shared-worker execution in PR #680). This is the sole
+// shared-exact threshold: a queue at or above this rate shards across every
+// eligible worker; a queue below it runs under a single owner.
+//
+// Evidence basis (#698):
+// - Drain-path userspace micro-bench `cos_exact_drain_throughput_micro_bench`
+//   (in `afxdp::tx::tests`, run with
+//   `cargo test --release -- --ignored --nocapture`; debug-build numbers are
+//   not meaningful for this baseline) measures the inner
+//   `drain_exact_local_fifo_items_to_scratch` +
+//   `settle_exact_local_fifo_submission` loop in isolation with setup work
+//   excluded from the timed region. Baseline on the development host is
+//   comfortably above MIN (order of a few Mpps / tens of Gbps at 1500 B);
+//   drain alone is not the limiter there.
+// - This bench only rules out the inner drain loop as the immediate
+//   limiter on the development host. It does NOT by itself validate MIN
+//   on other deployment hardware, and it does not fully attribute the
+//   remaining ceiling to non-drain work without a live single-worker
+//   measurement.
+// - The 2.5 Gbps figure is best read as a per-worker *aggregate* budget
+//   threshold consistent with the PR #680 collapse shape: there the drain
+//   loop failed to absorb 10g line-rate despite drain alone being able
+//   to go much faster, because non-drain per-packet work (RX, forwarding,
+//   NAT, session-lookup, conntrack) consumed the per-packet cycle budget
+//   that drain+completion needed to keep up.
+// - The ceiling is a property of the full per-worker pipeline, not of
+//   the interface shaper — it does not scale with iface rate.
+pub(super) const COS_SHARED_EXACT_MIN_RATE_BYTES: u64 = 2_500_000_000 / 8;
+
+
+
+
+
+
+
+
+
+/// #709: snapshot the owner-profile counter set from a `BindingLiveState`
+/// into a struct-local copy. Histograms are fixed-cap arrays on both
+/// sides; copying into an owned value lets the caller attribute the
+/// same snapshot to multiple queues without re-reading the atomics
+/// (which would tear across queues in the same scrape).
+pub(super) struct OwnerProfileSnapshot {
+    pub(super) drain_latency_hist: [u64; DRAIN_HIST_BUCKETS],
+    pub(super) drain_invocations: u64,
+    pub(super) drain_noop_invocations: u64,
+    pub(super) redirect_acquire_hist: [u64; DRAIN_HIST_BUCKETS],
+    pub(super) owner_pps: u64,
+    pub(super) peer_pps: u64,
+    /// #760 instrumentation, binding-scoped. Bytes delivered via
+    /// the post-CoS backup transmit paths in `drain_pending_tx`
+    /// — these never passed a queue's token gate. Surfaced on
+    /// the same "unambiguous owner-local exact queue" row the
+    /// other binding-scoped fields use.
+    pub(super) post_drain_backup_bytes: u64,
+    /// #760 instrumentation, binding-scoped. Bytes observed at the
+    /// three `apply_*` tx_bytes sites, incremented unconditionally.
+    /// Compare against the sum of per-queue `drain_sent_bytes`; any
+    /// gap is shaped traffic that bypassed the per-queue write via
+    /// an `apply_*` early-return.
+    pub(super) drain_sent_bytes_shaped_unconditional: u64,
+}
+
+
+
+
+
+
+
+pub(crate) fn push_recent_exception(
+    recent_exceptions: &mut VecDeque<ExceptionStatus>,
+    exception: ExceptionStatus,
+) {
+    if recent_exceptions.len() >= MAX_RECENT_EXCEPTIONS {
+        recent_exceptions.pop_front();
+    }
+    recent_exceptions.push_back(exception);
+}
+
+pub(crate) fn push_recent_session_delta(
+    recent_session_deltas: &mut VecDeque<SessionDeltaInfo>,
+    delta: SessionDeltaInfo,
+) {
+    if recent_session_deltas.len() >= MAX_RECENT_SESSION_DELTAS {
+        recent_session_deltas.pop_front();
+    }
+    recent_session_deltas.push_back(delta);
+}
+
+pub(crate) struct BindingLiveSnapshot {
+    pub(crate) bound: bool,
+    pub(crate) xsk_registered: bool,
+    pub(crate) xsk_bind_mode: String,
+    pub(crate) zero_copy: bool,
+    pub(crate) socket_fd: c_int,
+    pub(crate) socket_ifindex: i32,
+    pub(crate) socket_queue_id: u32,
+    pub(crate) socket_bind_flags: u32,
+    pub(crate) rx_packets: u64,
+    pub(crate) rx_bytes: u64,
+    pub(crate) rx_batches: u64,
+    pub(crate) rx_wakeups: u64,
+    pub(crate) metadata_packets: u64,
+    pub(crate) metadata_errors: u64,
+    pub(crate) validated_packets: u64,
+    pub(crate) validated_bytes: u64,
+    pub(crate) local_delivery_packets: u64,
+    pub(crate) forward_candidate_packets: u64,
+    pub(crate) route_miss_packets: u64,
+    pub(crate) neighbor_miss_packets: u64,
+    pub(crate) discard_route_packets: u64,
+    pub(crate) next_table_packets: u64,
+    pub(crate) exception_packets: u64,
+    pub(crate) config_gen_mismatches: u64,
+    pub(crate) fib_gen_mismatches: u64,
+    pub(crate) unsupported_packets: u64,
+    pub(crate) flow_cache_hits: u64,
+    pub(crate) flow_cache_misses: u64,
+    pub(crate) flow_cache_evictions: u64,
+    pub(crate) flow_cache_collision_evictions: u64,
+    /// #941 Work item D: count of V_min hard-cap activations on this
+    /// binding (per `update_binding_debug_state` flush of each queue's
+    /// scratch counter). Acceptance gate: under normal load, the
+    /// override-rate (this / `drain_invocations` aggregated across
+    /// queues) stays below 5 %.
+    pub(crate) v_min_throttle_hard_cap_overrides: u64,
+    pub(crate) session_hits: u64,
+    pub(crate) session_misses: u64,
+    pub(crate) session_creates: u64,
+    pub(crate) session_expires: u64,
+    pub(crate) session_delta_pending: u64,
+    pub(crate) session_delta_generated: u64,
+    pub(crate) session_delta_dropped: u64,
+    pub(crate) session_delta_drained: u64,
+    pub(crate) policy_denied_packets: u64,
+    pub(crate) screen_drops: u64,
+    pub(crate) snat_packets: u64,
+    pub(crate) dnat_packets: u64,
+    pub(crate) slow_path_packets: u64,
+    pub(crate) slow_path_bytes: u64,
+    pub(crate) slow_path_local_delivery_packets: u64,
+    pub(crate) slow_path_missing_neighbor_packets: u64,
+    pub(crate) slow_path_no_route_packets: u64,
+    pub(crate) slow_path_next_table_packets: u64,
+    pub(crate) slow_path_forward_build_packets: u64,
+    pub(crate) slow_path_drops: u64,
+    pub(crate) slow_path_rate_limited: u64,
+    pub(crate) kernel_rx_dropped: u64,
+    pub(crate) kernel_rx_invalid_descs: u64,
+    pub(crate) tx_packets: u64,
+    pub(crate) tx_bytes: u64,
+    pub(crate) tx_completions: u64,
+    pub(crate) tx_errors: u64,
+    pub(crate) redirect_inbox_overflow_drops: u64,
+    pub(crate) pending_tx_local_overflow_drops: u64,
+    pub(crate) tx_submit_error_drops: u64,
+    // #760 triage: surfaced on BindingStatus so operators can
+    // compare binding-level vs per-queue drain accounting.
+    pub(crate) post_drain_backup_bytes: u64,
+    pub(crate) drain_sent_bytes_shaped_unconditional: u64,
+    // #760 (PR #773): drop-filter counters for CoS-bound items
+    // that reached the post-CoS backup paths. Non-zero indicates
+    // a cross-worker routing failure the bounded ingest-drain
+    // loop did not absorb.
+    pub(crate) post_drain_backup_cos_drops: u64,
+    pub(crate) post_drain_backup_cos_drop_bytes: u64,
+    // #710: `no_owner_binding_drops` is intentionally NOT snapshotted
+    // per-binding. The atomic on `BindingLiveState` accumulates drops
+    // for mechanical accounting (the increment site can only write to
+    // `bindings.first_mut()`), but the operator-facing aggregate lives
+    // at `ProcessStatus::cos_no_owner_binding_drops_total`, summed
+    // across every live state by
+    // `Coordinator::cos_no_owner_binding_drops_total()`.
+    pub(crate) direct_tx_packets: u64,
+    pub(crate) copy_tx_packets: u64,
+    pub(crate) in_place_tx_packets: u64,
+    pub(crate) direct_tx_no_frame_fallback_packets: u64,
+    pub(crate) direct_tx_build_fallback_packets: u64,
+    pub(crate) direct_tx_disallowed_fallback_packets: u64,
+    pub(crate) debug_pending_fill_frames: u32,
+    #[allow(dead_code)]
+    pub(crate) debug_spare_fill_frames: u32,
+    pub(crate) debug_free_tx_frames: u32,
+    pub(crate) debug_pending_tx_prepared: u32,
+    pub(crate) debug_pending_tx_local: u32,
+    pub(crate) debug_outstanding_tx: u32,
+    pub(crate) debug_in_flight_recycles: u32,
+    /// #878: per-binding UMEM total frames (set once at worker
+    /// construction). Used as the denominator for the `show chassis
+    /// forwarding` Buffer% display; numerator comes from
+    /// `umem_inflight_frames` published once per second by the
+    /// owning worker.
+    pub(crate) umem_total_frames: u32,
+    /// #878: configured TX-ring depth (set once at worker
+    /// construction). `outstanding_tx / tx_ring_capacity` is the
+    /// second pressure signal aggregated by Buffer%.
+    pub(crate) tx_ring_capacity: u32,
+    /// #878: UMEM in-flight gauge published in a single store from
+    /// the worker's per-second debug tick — no torn-load risk on
+    /// the read side.
+    pub(crate) umem_inflight_frames: u32,
+    // #802: ring-pressure snapshot fields. Mirrored from BindingLiveState
+    // atomics that are published by the worker's per-second debug tick.
+    pub(crate) dbg_tx_ring_full: u64,
+    pub(crate) dbg_sendto_enobufs: u64,
+    // #802/#804: split — see `BindingLiveState` for write-site semantics.
+    pub(crate) dbg_bound_pending_overflow: u64,
+    pub(crate) dbg_cos_queue_overflow: u64,
+    pub(crate) rx_fill_ring_empty_descs: u64,
+    pub(crate) last_heartbeat: Option<chrono::DateTime<Utc>>,
+    pub(crate) last_error: String,
+    // #709: owner-profile telemetry snapshot. Fixed-size arrays (no
+    // `Vec`) to keep the snapshot allocation-free on the hot path;
+    // readers that want a `Vec` for JSON can copy on demand.
+    pub(crate) drain_latency_hist: [u64; DRAIN_HIST_BUCKETS],
+    pub(crate) drain_invocations: u64,
+    pub(crate) drain_noop_invocations: u64,
+    pub(crate) redirect_acquire_hist: [u64; DRAIN_HIST_BUCKETS],
+    pub(crate) owner_pps: u64,
+    pub(crate) peer_pps: u64,
+    /// #812: per-queue TX submit→completion latency histogram +
+    /// count + sum-ns. Fixed-size array (same pattern as
+    /// `drain_latency_hist`). The array is materialized into a
+    /// `Vec<u64>` only at the JSON/protocol boundary; the snapshot
+    /// itself stays allocation-free.
+    pub(crate) tx_submit_latency_hist: [u64; TX_SUBMIT_LAT_BUCKETS],
+    pub(crate) tx_submit_latency_count: u64,
+    pub(crate) tx_submit_latency_sum_ns: u64,
+    /// #825: per-kick `sendto` latency histogram + count +
+    /// sum-ns + EAGAIN-retry count. Fixed-size array matches
+    /// `tx_submit_latency_hist`; materialized into a `Vec<u64>`
+    /// at the JSON/protocol boundary, the snapshot itself stays
+    /// allocation-free.
+    pub(crate) tx_kick_latency_hist: [u64; TX_SUBMIT_LAT_BUCKETS],
+    pub(crate) tx_kick_latency_count: u64,
+    pub(crate) tx_kick_latency_sum_ns: u64,
+    pub(crate) tx_kick_retry_count: u64,
+}
